Revamp TypeScript usage throughout the codebase
Closes #496
