--- check/check-body.js
@@ -1,2 +1,0 @@
-const check = require('./check');
-module.exports = (fields, message) => check(fields, ['body'], message);
