+++ check/index.js
@@ -1,8 +1,12 @@
-exports.check = require('./check-all');
-exports.body = require('./check-body');
-exports.cookie = require('./check-cookies');
-exports.header = require('./check-headers');
-exports.param = require('./check-params');
-exports.query = require('./check-query');
-exports.validationResult = require('./validation-result');
-exports.oneOf = require('./one-of');
+const { check, body, cookie, header, param, query } = require('./validation-chain-builders');
+
+module.exports = {
+  check,
+  body,
+  cookie,
+  header,
+  param,
+  query,
+  oneOf: require('./one-of'),
+  validationResult: require('./validation-result')
+};
