+++ check/index.spec.ts
@@ -1,0 +1,27 @@
+import {
+  // check.d.ts
+  CustomValidator,
+  Validator,
+  ValidationChain,
+  ValidatorOptions,
+
+  // validation-chain-builders.d.ts
+  ValidationChainBuilder,
+  buildCheckFunction,
+  check,
+  body,
+  cookie,
+  header,
+  param,
+  query,
+
+  // oneOf.d.ts
+  oneOf,
+
+  // validation-result.d.ts
+  validationResult,
+  ErrorFormatter,
+  Result,
+  ResultFactory,
+  ResultFactoryBuilderOptions,
+} from './index';
