+++ check/location.d.ts
@@ -1,0 +1,1 @@
+export type Location = 'body' | 'params' | 'query' | 'headers' | 'cookies';
