+++ check/validation-result.d.ts
@@ -1,15 +1,28 @@
 import * as express from 'express';
-import { Result, ErrorFormatter } from '../shared-typings';
+import { Location } from './location';
 
-export const validationResult: validationResult;
+export type ErrorFormatter<T = any> = (error: {
+  location: Location,
+  param: string,
+  msg: any,
+  value: any
+}) => T;
 
-export interface validationResult {
-  (req: express.Request): Result;
-  withDefaults(options: WithDefaultsOptions) : this;
+export interface Result<T = any> {
+  array(options?: { onlyFirstError?: boolean }): T[];
+  formatWith(formatter: ErrorFormatter<T>): this;
+  isEmpty(): boolean;
+  mapped(): Record<string, T>;
+  throw(): void;
 }
 
-interface WithDefaultsOptions {
+export const validationResult: ResultFactory;
+
+export interface ResultFactory {
+  <T>(req: express.Request): Result<T>;
+  withDefaults(options?: Partial<ResultFactoryBuilderOptions>) : this;
+}
+
+interface ResultFactoryBuilderOptions {
   formatter: ErrorFormatter;
 }
-
-export default validationResult;
