+++ filter/index.js
@@ -1,6 +1,16 @@
-exports.matchedData = require('./matched-data');
-exports.sanitize = require('./sanitize-all');
-exports.sanitizeBody = require('./sanitize-body');
-exports.sanitizeCookie = require('./sanitize-cookies');
-exports.sanitizeParam = require('./sanitize-params');
-exports.sanitizeQuery = require('./sanitize-query');
+const {
+  sanitize,
+  sanitizeBody,
+  sanitizeCookie,
+  sanitizeParam,
+  sanitizeQuery
+} = require('./sanitization-chain-builders');
+
+module.exports = {
+  sanitize,
+  sanitizeBody,
+  sanitizeCookie,
+  sanitizeParam,
+  sanitizeQuery,
+  matchedData: require('./matched-data')
+};
