+++ filter/index.spec.ts
@@ -1,0 +1,17 @@
+import {
+  // matched-data.d.ts
+  matchedData,
+  MatchedDataOptions,
+
+  // sanitize.d.ts
+  Sanitizer,
+  SanitizationChain,
+
+  // sanitization-chain-builders.d.ts
+  SanitizationChainBuilder,
+  sanitize,
+  sanitizeBody,
+  sanitizeCookie,
+  sanitizeParam,
+  sanitizeQuery,
+} from './index';
