+++ filter/sanitization-chain-builders.d.ts
@@ -1,0 +1,13 @@
+import { Location } from '../check/location';
+import { SanitizationChain } from './sanitize';
+
+interface SanitizationChainBuilder {
+  (fields: string | string[], message?: any): SanitizationChain;
+}
+
+export function buildSanitizeFunction(location: Location[]): SanitizationChainBuilder;
+export const sanitize: SanitizationChainBuilder;
+export const sanitizeBody: SanitizationChainBuilder;
+export const sanitizeCookie: SanitizationChainBuilder;
+export const sanitizeParam: SanitizationChainBuilder;
+export const sanitizeQuery: SanitizationChainBuilder;
