--- filter/sanitize-cookies.js
@@ -1,2 +1,0 @@
-const sanitize = require('./sanitize');
-module.exports = fields => sanitize(fields, ['cookies']);
